backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-cloud-speech==2.23.0
google-cloud-aiplatform==1.43.0
google-cloud-firestore==2.14.0
elevenlabs>=1.0.0
python-multipart==0.0.6
python-dotenv==1.0.1
